add examples

+++ level_set_machine_learning/core/temporary_data_handler.py
@@ -49,7 +49,8 @@ class TemporaryDataHandler:
     def remove_tmp_data(self):
         """ Delete the temporary data folder and contents
         """
-        shutil.rmtree(self.tmp_data_location)
+        if os.path.exists(self.tmp_data_location):
+            shutil.rmtree(self.tmp_data_location)
 
     def _get_h5_file_path(self):
         """ Returns the full path to the hdf5 file holding the temporary data
